Apply review feedback
